Use daybreak/nightfall for wallpaper instead of sunrise/sunset
This is because it makes more sense to say when daytime and nighttime
start, than when daytime starts and ends. (At least in my head it does).
¯\_(ツ)_/¯

Simplify the tests to use a single WALLPAPER constant, so that the
calculations all reflect the same thing. This follows the dynamic Mojave
wallpapers that come with macOS Mojave.