Refactor get_image to split out calculations per time period
Instead of trying to be clever and combine all the calculations into
one, split it out so that each time period (before sunset, daytime,
after sunset) is handled separately.

We now make the assumption that images to be used during daylight hours
are first, and nighttime images come after. In other words, we don't
assume the list wraps around at some arbitrary point anymore; only when
it reaches the last night image. This removes the need for an `offset`
value (technically, it's now the number of daytime images).