Use NonZeroU32 for Wallpaper config
Implement FromStr for Config, and call it from the TryFrom<PathBuf>
impl.

This simplifies a lot of the checks, by ensuring at compile time that
`wallpaper.day_images` and `wallpaper.night_images` are not zero.

I guess as a downside, it complicates the tests a bit, since you can't
call non-constant methods to assign a `const`, so we use `lazy_static`.